New Toppings and Small Fixes
I added mushrooms, olives and onions.
I remodeled the sausage model and lowered the vertices of the cheese model by 46 vertices I think.
I improved on the pizza.cs script to include the topping data as arrays to avoid hardcoding its features and make it easier to add new toppings.