Improved Pizza Script
I improved on the pizza script to make it check the distance between toppings and check which side it is to ensure that the player is putting them correctly and not spamming them in random places.

[Plans]
I plant to add the furnace that GPGP uses and add its mechanics, though it may be hard as it is a 3D game and not a 2D game, but I pulled through this much so there is nothing stopping me now.