Finished Cooking System
I finally finished the whole cooking system from the making of the pizza to the cooking to the slicing to the packaging, The only thing left for the game to be ready fundementally is adding a customer system and maybe a money system. instead of the customer giving you exactly the mmoney, he gives you money and oyu give him change by using the recipe book! (maybe annoying but fun)