Encryption and Dialog system
I've added a dialog system between you and the customer which will be the backbone for orders.
the encryption system I've implemented is for use in a story in my mind, where you gather paper and encryption keys and try the keys to decrypt each paper which is written on it gibberish to actual lines which will help you make the ancient pizza.
Inspired by chapter 5 of GPGP.

I am planning to add the customer system and money system to finish the pre-beta stage if you will (between alpha and beta) in the next commit.