Customer system completed
What a complicated script i had to type for this customer system.
Anyway the customer system is now ready and behaves pretty closely to the actual game's customer.
next time I'll add its patience, money counter and make ingredients cost something when placed and some visuals when the customer pays something, and add L and R symbolizing Left and Right respectively to help the player put ingredients in the right sides as the customer know exactly which side of the pizza it is.