Huge changes
I made some huge changes and one of them is Customer System overhaul.
I made the customer not be picky about which side he wants his pizza much like gpgp, Though it made the code more complex it worked in the end.

I also fixed some bugs along the way like the camera locking in place when doing certain actions with the right button, right clicking the dough to change to pizza when picking it made the player locked so that was fixed, and the customer now speaks normally and not go gibberish if you skip his talk by asking what? or hint.