Added Preview system
I made a rather nice quality of life feature, A preview system!
What it does is it shows a very transparent version of the topping you wish to add on your pizza before you actually click the left mouse button, this ensures that you don't make any mistakes when putting your toppings or wanna put them in a very specific space in the pizza. I also made the customer come with randomized orders instead of coming with a pre-selected one for debugging.