Stricter Customer
I made the customer script more strict towards your toppings and I added a hatedToppings list like how a customer may ask for NO onions, and If you add even one of them;  They'll refund it all up! I also added a maximumtoppingsallowed, this is to allow other toppings to appear on the pizza, I may add a disgusted face with no tip if it almost reaches the max, maybe another variable for that. and a hated face when you reach the topping maximum. I'll also think of making better lines that correspond to the issue you did rather than general ones if you failed the CheckIngrediants() function.