Objective script complete
I finished the objective script, I plan to add a story that challenges each topping one by one before you get the other toppings to make a supreme pizza.
I also made the supply visual that is similar to how gpgp shows it to you, But it is with every topping bought from the merchant.
The items finally have visuals, And I plan to add a next day and leave and some graphics to finish a bit of the settings