Tutorial
Added a small tutorial to teach the basics for the player, It is played automatically the first time you enter, then you can choose whenever you want the tutorial before you start the game if you forgot the controls.
One bug it the animation bug of the Tutorial button which I should fix soon.
I might add those hint controls used in games but that's a side feature so it is pushed for later.